+++ llvm/tools/llubi/lib/CMakeLists.txt
@@ -7,6 +7,7 @@ set(LLVM_LINK_COMPONENTS
 add_llvm_library(LLVMUBAwareInterpreter
   STATIC
   Context.cpp
+  ExecutorBase.cpp
   Interpreter.cpp
   Value.cpp
   )
